*Added random times to mole *Added HUD to game **Issue with whack percent not updating due to non-hit

+++ Striker/Assets/Scripts/Player/Player.cs
@@ -35,7 +35,7 @@ namespace Assets.Scripts.Player
         {
             get
             {
-                return (float)(Whacks / WhackAttempts);
+                return (float)Whacks / (float)WhackAttempts;
             }
         }
